Add new monsters from NetHack 3.7.0.
From NetHack 3.7.0 (git commits 7817e69, c63165e, and 16562b2). The
method of monster displacement that already exists in EvilHack was kept,
I feel that implementation (from SpliceHack) is better than what was
developed for vanilla NetHack. Otherwise it's all the same.

Also in this commit, a couple bits for the new amulet tiles I missed
from the last commit.