More tweaks, fixes, and changes with Infidel role.
A few more adjustments to the Infidel role:

* Corrected some grammar mistakes with the quest text.
* Changed how the Infidel is addressed upon ascension - instead of demon
lord/lady, it's 'Archfiend of Moloch' (if crowned) or 'Emissary of
Moloch' (if not crowned), both are independent of gender.
* Fixed (with the Amulet), changed it to (with the Idol) for ascending
Infidels.
* Changed the name of the quest leader, from 'preacher of Moloch' to
'Magistra of Moloch' and specified female gender. This naming convention
follows the hierarchy of the Church of Satan; their current
chairmistress is referred to as 'Magistra'. The title 'preacher' makes
me think of a small town pastor somewhere in the bible belt preaching to
their flock... had to be changed.
* Add PM_DEMON to is_placeholder()
* Removed the magic marker from the High Priest of Moloch's inventory
(sanctum) and placed it within a crystal chest right next to the altar
instead. Reason being - since that marker is potentially the only one
that can ever be found in the game, having it in the priests possession
effectively screws over Infidels, since that priest is peaceful towards
them, and killing a co-aligned priest is frowned upon.
* Removed several alignment/luck penalties for Infidels killing peaceful
types. Killing pets or a co-aligned priest is still bad... don't do it.

At this point, the changes to the Infidel role has strayed far enough
from the original patch, that will need to be made clear on the
NetHackWiki.