Mummies have a withering attack.
From SpliceHack (modified). Instead of mummies dealing a poison-based
claw attack, they can now cause its target to 'wither away' on a
successful hit. Withering away means that for a random amount of turns,
the target that is afflicted loses a small amount of hit points each
turn, and cannot regenerate hit points back until the withering effect
wears off. If the target is already low on hit points, this can prove
fatal. Non-living monsters are immune to this effect. The player with
see a 'Wither' status on the bottom line if so afflicted. This effect
lasts slightly longer than it does in SpliceHack, but the player (and
monsters) have an extra means at their disposal to cure it. Most of the
same methods that the player can use to cure an illness, also works for
withering away.

Also in this commit, noticed and fixed a couple bottom line displays for
other custom notices.