Allow giant race to play as Samurai, tweak to the Tsurugi of Muramasa.
Some changes for the Samurai role:

* Giants can now play as a Samurai.
* The Tsurugi of Muramasa has been spiced up a bit to garner more
interest in using it.

First bit - giants can play as Samurai. Great. But what about their suit
of armor? That's kind of a hallmark of any samurai. So for giant samurai
*only*, they start with a set of 'large splint mail'. New object whose
stats closely match that of regular splint mail, except that it will
never spawn randomly, and its weight is nearly double that of regular
splint mail (500 aum vs 300 aum). This is the only piece of body armor a
giant can wear, and only giant samurai. If any other role finds or
wishes for this particular armor, they won't be able to wear it.
Implementing this felt a bit like a kludge. May revisit this later when
it's not almost 2am in the monring.

Second bit - I added half-physical damage to the Tsurugi to give
incentive to players so they'd want to actually use it. Almost everyone
that plays a samurai will ditch their quest artifact and dual-wield
Excalibur along with their katana. I figured half-physical damage would
be good because 1) it's half-physical damage, which is awesome and 2) no
lawful artifacts have that. This might be too OP, especially for a giant
samurai, but it sounds fun, so lets give it a try and see how it goes.