Even more Infidel role changes.
We're getting there. After conferring the Tomsod (the Infidel role
author) as well as one or two folks who have been play-testing this,
here are the latest changes:

* The Infidel quest leader no longer has a set gender (they can spawn as
either male or female), and their title has changed again to be more
gender neutral.
* Some very minor tweaks with the quest text (found a case of present
tense which should have been past tense).
* Infidels no longer start with trident skill. Instead, trident skill is
unlocked once they become crowned, and can be trained up to expert
skill.
* Illithid Infidels still retained the use of their 'psionic wave'
ability (technically a spell) after being crowned. This has been
suppressed as this is supposed to be an illithid-only ability.
* Title change - once an Infidel is crowned, they then become an
'Emissary of Moloch' and achieve the status of 'Archfiend' upon
ascension. This sounds a lot better.
* Spellcasting changes with or without the Amulet or imbued Idol again.
Spell power drain will always be present while holding the Amulet or the
imbued Idol now. Because the Infidel role will have to endure this from
start to finish (as the role author envisioned), the spell power drain
for them has been lessed to an extent. Also (and this is the
controversial change), Infidels will not be able to cast spells at all
if they do not have the Amulet or imbued Idol in their open inventory.
Reason - incentive to keep these objects on them at all times. They are
needed to win, the theme of the entire quest revolves around the premise
that the player has the Amulet in their possession, and it prevents the
player from just ditching the Amulet under a boulder on dungeon level
one and not messing with it again until they're ready to enter the
Sanctum (that 'strategy' completely undermines the theme of this role).
We'll see how this particular change is received, but after much
discussion and thought, I think this fits the narrative.