Reverting spell power drain for Infidels; minor altar sacrifice tweaks.
Infidels now go back to the way I had this initially - while they have
the amulet, spell power is drained as they cast spells, just like for
every other role. Once they sacrifice the amulet and imbue the Idol,
Moloch has control of the Amulet's power now imbued in his Idol, and
suppresses the spell power drain effect, allowing the Infidel to realize
their full potential. Not having the Amulet in the Infidels possession
still blocks the ability to cast at all, that's still in (player can
drop the Idol no problem and still cast), along with teleportation
dampening effect. Not changing this anymore - this is how things will be
when this version goes live, and we can make adjustments after the fact
after people have had a real chance to play and try out this role.

Also in this commit - I've tagged Infidels as primary spellcasters for
the purposes of altar sacrificing, so the regular gear they receive will
be more inline with what a wizard would get. Couple other minor tweaks
in there as well (non-Infidel related).