Infidel ascension feedback correction, power regen tweak.
Missed a piece in artifact.c regarding the status title given to an
Infidel upon ascension. In light of the harsh spell power drain during
the early game for an Infidel that does the right thing by holding on to
the Amulet, let's throw them a bone - their regular power regeneration
is now on par of that as a wizard.