New artifacts: Angelslayer, Bag of the Hesperides, Butcher, Wand of Orcus.
Some new artifacts have been added to the game:

* Angelslayer - an artifact trident that warns of angels, confers
searching and half-spell damage, and has a fire-based attack. Not a true
bane weapon, but angels are susceptible to fire and having half-spell
damage helps against their magic missile attack.
* Bag of the Hesperides - this is the bag given to Perseus by the
Hesperides (nymphs) to aid him in his quest to defeat Medusa. It was
used to store and carry Medusa's head. The bag's base type is a bag of
holding, confers protection (one level of MC), is made of dragonhide,
and greatly reduces the weight of items placed in it, even moreso than a
regular bag of holding. It also repels water just like an oilskin sack.
Can be found with the captive pegasus at the end of the Ice Queen's
branch. Cannot be wished for, but as it is guaranteed to be in the game,
it will not count against the player for artifact wishing purposes
(number of artifacts generated).
* Butcher - this massive triple-headed flail belongs to the demon lord
Yeenoghu. Legend states that Butcher was created from the thighbone and
skin of a god Yeenoghu killed. The base item is a triple-headed flail
(new object) which is so large it requires two hands to wield if
human-sized, and its damage output is greater than even a dwarvish
mattock. Butcher has a stun attack, much like what the artifact
Magicbane can do, but unlike Magicbane, Butcher does not also cause fear
or cancellation. Yeenoghu spawns with this artifact and will use it.
Cannot be wished for, and will also not count against the player for
artifact wishing purposes.
* Wand of Orcus - the infamous weapon of the demon prince Orcus, its
base item is a heavy mace (new object) which is made of obsidian and
topped with a huge skull. Appears in Orcus' inventory as he spawns. Any
living creature struck by this artifact will either have their maximum
hit points drained, or if not magic resistant, a small chance of being
killed outright. The base item delivers a fair amount of damage in its
own right - wielded by Orcus, its damage output is significant. The Wand
of Orcus can never be used by the player, any attempt to pick it up
will be unsuccessful, and will harm the player in the attempt. Cannot
be wished for, and will also not count against the player for artifact
wishing purposes.