New level (nymph) and new special room (garden).
Addition of a new level that can randomly spawn between levels 5 and 15
- nymph level: the entire level is controlled by several nymphs, and
their leader is Aphrodite, a unique nymph that is stronger and faster
than the rest. Most rooms on this level have a chance for trees to
spawn, and the largest room is filled with chests and loot, a few trees,
and a fountain.

New room type - a garden. Also nymph-themed, can start to appear after
level 7, gardens are green in color and have a random number of nymphs,
trees and fountains. Various sounds can be heard on any level this rooms
spawns on, giving the player a hint that a garden is there.

Both of these were pulled from UnNetHack, and the nymph level in
UnNetHack originally came from Slash'EM.