Extended achievement and conduct fields for xlogfile.
From UnNetHack commit 4e8a69e, modified. This adds human readable data
to the xlogfile for various achievements, events and conducts. This in
turn makes it very easy for any type of scoreboard/tournament system to
track these, while not having to worry about bitfields changing for
those particular events in the future.

Also in this commit, a slight format edit to a new monster from an
earlier commit.