Armor made of a hated material deflecting a monsters attack.
Quite awhile ago I had added in Slash'EM's 'descriptive miss' messages.
What that does, is if an attacker misses you, the feedback receives is
more verbose than just 'The foo misses!'. Depending on the die roll, it
could be a narrow miss, or maybe you dodged or evades the attack. If
you're wearing armor, there's a chance that piece of armor deflected the
attack.

In this commit, if a piece of armor blocked an attack, and the material
of that piece of armor causes damage to the attacker if it comes in
contact with their skin, AND if the attacker attacks bare-handed and
isn't wearing gloves, they can take a small amount of sear damage from
the contact.

At some point this needs to be the same way for players attacking a
monster bare-handed without gloves.