Descriptive miss messages (player vs monster, monster vs monster).
Lifted this from Slash'EM (modified). Figured if there are descriptive
miss messages for monsters attacking the player, it should exist for the
other two scenarios as well. I haven't added any handling for hated
materials and armor deflecting bare-handed attacks effects for that yet,
I'll save that for a later time.