Fix: giant samurai breaking out of their large armor when poly'd into a large form.
The large splint mail that giant samurai start with is made specifically
for huge humanoids. Being poly'd into something larger than human-sized
shouldn't break it. If poly'd into something MZ_LARGE or larger, if the
form is humanoid, nothing happens (armor is still worn). If the form is
not humanoid, the armor simply falls off.