The Wand of Orcus is always known.
It didn't feel as scary being attacked by Orcus as he wielded his
'gemstone heavy mace named Wand of Orcus', and the player will never be
able to wield it or even handle it in inventory, so they'd never be able
to ID it. So make it universally known - makes for a better experience.