Fix: proper handling of monster's weapon/gloves when attacking a black dragon.
For quite some time now, if the player attacked a baby black
dragon/black dragon, wielded objects and worn gloves would be
disintegrated first when landing a hit, and would only disintegrate the
player themselves if they attacked bare-handed. But if a monster
attacked a black dragon, or if a monster attacked the player poly'd into
a black dragon, it never mattered if the monster was wielding a weapon
or wearing gloves - they would be disintegrated outright if they landed
a hit and the RNG blessed the player with a good roll. This commit
addresses that shortcoming. Now, for both cases just described, a
monster's wielded weapon or worn gloves would be disintegrated first.

Also in this commit, is a fix for an oversight on my part in regards to
monsters attacking black dragons. The odds of being disintegrated upon
landing a hit are supposed to be 1 in 20 (expressed as !rn2(20)), but
for testing purposes I had set that for rn2(20) so the odds would be 19
in 20. Forgot to set that back, which would explain why having a herd of
tame black dragons had become super popular of late. Oops.