Some tweaks to random vaults.
First, a quick adjustment to the 'massacre' room - convicts and infidels
weren't being represented. I also switched the corpses back to regular
player corpses and not player monster corpses.

Second - there are two types of rooms known as 'splits', one vertical,
the other horizontal. The room can be split by either pools of lava or
iron bars. Most times there's an opening available between the two room
sections, but there is a small chance that there won't be. If stairs or
some other desirebale object is on one side of the split without an
opening, the player will have to be resourceful in finding a way to the
other side. That in and of itself isn't the problem - the problem is
when this type of room spawns with no opening on dungeon level 2 and the
stairs spawn on one side of the split. That early in the game, most
players won't have the resources available to them to figure out a way
around or through. I've set the minimum level depth for these rooms at
dungeon level 5 - this way the stairs down into the gnomish mines won't
ever spawn inside one of these rooms, and this gives the player ample
opportunity to discover various items needed to deal with these rooms
deeper in the dungeon.