Allow skilled attack spell casters to cast 'basic' level for various attack spells.
From xNetHack commit 6b357f7, which got it from UnNetHack, which got it
from dNetHack, which got it from...

This affects the spells fireball, cone of cold, poison blast and acid
blast - if the caster is skilled or expert in attack spells, casting
these spells prompts 'Cast advanced spell?' giving the caster the option
for the advanced effect (area of effect) or the basic effect (beam on a
single point). If basic effect is chosen, only half of the spell power
needed to cast it is used.