Fix: prevent potential segfault using a stethoscope to listen to eggs.
A few days ago, aosdict pointed out that listening to a NON_PM egg and
said egg not being stale could result in a segfault. This commit
addresses that. Recently in xNetHack there is a more robust method of
listening to eggs to determine what's inside, may pull that in at a
later time. For now what I have here is fine.