Player monsters that spawn in the endgame use names from the topten list.
One side effect of this is that if you have the same player in the
topten multiple times, the odds go up that their name will be chosen
more than once. Perhaps there's a way to address that, so once a
particular name is chosen, don't choose it a second time. If the topten
list is empty, the game will pick from the developers name array
instead.