Fix: Gjallar not affecting various monsters predisposed to wait.
The artifact horn Gjallar was designed to be the loudest object in the
game, even more effective than a drum of earthquake. It would do as it
claimed, waking up any sleeping monster anywhere on the map, regardless
of the players position. However, any monster with any kind of waitmask
flag would still be unaffected unless the player got within a certain
distance (for example, the maximum distance away one can be while trying
to awaken the Master Assassin in the rogue quest). This commit lifts
that restriction. Gjallar is so powerful, *any* creature will take
notice. This can have consequences, such as drawing attention to a
unique monster before the player was ready to do so (if this happens,
the player is made aware of it).