Increase starting pets survivability versus zombies.
Zombies are slow and mindless. All they want to do is to eat brains, but
smaller creatures are harder for them to catch. If a monster is MZ_TINY
or MZ_SMALL, they have a 66% chance of evading a zombies bite attack.
Since all starting pets (except ponies) fall within this size range,
their survivability versus zombies is increased. This affects the player
as well if they happen to be polymorphed into a small or tiny creature
(and not encumbered), or if they themselves are polymorphed into a
zombie and are trying to attack tiny or small monsters.

I also discovered and corrected a couple bugs with the bite attack in
regards to zombie vs monster.

This should preclude the necessity of little doggo bobble hats.