Prevent monsters zero-turn equipping and equipping with player nearby.
From xNetHack commit 8c4af50 - notes from that commit:

'There were two problems here, both exemplified by the scenario where the
player is fighting a battalion of soldiers through a choke point. What
happens here is that when each soldier dies and drops his armor onto a
pile of armor from previous kills, the next soldier steps forward, finds
a good (according to the monster AI) piece of armor, picks it up, and
starts equipping it, completely oblivious of the hero who is still
standing right there hitting them with a sharp object.

Problem 1: monsters can step forward, pick up stuff, and immediately
start equipping the stuff on that same turn.

Problem 2: monster AI prioritizes equipping with superior equipment and
disregards whether they're under any sort of threat.

This commit solves both problems. Problem 1 is solved by replacing all
instances of zero-turn monster equipping (calls to m_dowear that weren't
related to initialization of a new monster, the notable case being
immediately after picking up stuff), instead having them set a special
flag that indicates they should try to reassess their gear at the start
of their next turn.  To this end, I added a new function which wraps up
that behavior in a nice name.

Then, problem 2 is solved by implementing a condition that hostile
monsters must be outside a radius of 3 from where they think the player
is in order to choose to re-equip. If they're within this radius, the
flag telling them to re-equip remains set but they are not allowed to
choose it, meaning that as soon as they leave that radius they'll
re-equip. Note that monster knowledge of the player doesn't appear to be
updated at this point, so it is based on where the player started
instead of where they moved to; thus you can step into a space three
squares away from a monster (having moved closer) and it can still
choose to re-equip.

The radius of 3 is arbitrary; I initially had it at 5 but that seemed a
bit too restrictive on the monsters.'

Good stuff. I may play around with the radius value, as during testing
even a radius of 3 sometimes felt to be too much.