Fix: Sokoban zoo entrance door could become sealed shut while retrieving the prize.
If the sokoban zoo entrance door was closed at the same time the player
opened one of the three prize closet doors, it would be sealed and
replaced with an impassable wall just like the other two remaining
sokoban prize closet doors. This could permanently trap the player
inside the sokoban zoo unless they had a way to teleport out.