Fix: pet location inconsistency.
From NetHack 3.7 (git commit 1501924). The issue: if you had a tame
monster that was confused and could engulf, if it engulfed you and then
expelled you, there was a good chance that a bogus 'clone' of your pet
would be made, but was in fact considered the tail of your pet. It just
got more weird from that point on, and under certain scenarios would
lead to the game crashing. This bug wasn't fixed in vanilla NetHack
until version 3.7.0-16, and was reported to me recently from a
player using the windows binary. Pulling it in to EvilHack now as 1)
this variant has several more engulf-type monsters than vanilla, and 2)
I don't plan on merging 3.7 anytime soon.

Notes from the original commit:

'... if the pet attacked hero or another monster by eg.
swallowing them, the pet's location might've changed
during that attack. Count it as movement, so return
immediately.'