Fix: prevent shape-changing pets from reverting back to their original form while wearing a ring of protection from shape changers.
This 'bug' actually exists in vanilla NetHack, but only in regards to
chameleons, doppelgangers, werefoo and sandestins. In EvilHack it also
affected vampire types. Not sure if this is a bug from vanilla NetHack
or if it's intended behavior. Regardless, we're addressing it here. If
the player has invested a lot of time and effort growing up their pet,
only to have it revert back to its base form just because said player
put on a ring intended to protect them for hostile shape changers...
seems like a bug to me.