Fix: strength attribute abuse (giants).
This has been an issue for quite some time now - weird strength
attribute loss, most often as a giant. This looks as if this was
self-induced when I added giants as a playable race about a year and a
half ago. Chalking this up to a bad merge, not sure why this was the way
it was. Anyways... ran a few tests, strength attribute gain appears to
be back to normal. More testing for confirmation wouldn't be a bad
thing.

Bit of code formatting cleanup in this commit also.