Skip Elbereth learning check when creating graves
Because creating a grave when the hero dies involves
make_engr_at(engrave.c), the check for having learned Elbereth is
applied when a grave would have Elbereth on it (e.g. because the hero
was killed by a monster she had named Elbereth, or because the hero
was named Elbereth). This means that post-death the player receives a
message about not knowing the word, and the actual engraving on the
headstone in the bones file is replaced in its entirety with the
selected Elbereth replacement word.

Skipping this test while creating headstones means this won't happen,
and graves containing Elbereth will be treated like any other grave.