Fix: possible to destroy the Amulet and the invocation items in sewage.
Discovered a rather serious mistake in regards to how objects behave
when trying to dip them into sewage. It was possible to destroy the
Amulet or any of the invocation items by dipping them in sewage. This
commit addresses that oversight. I also cleaned up a bit of code in a
couple other places, using obj_resists() instead of reinventing the
proverbial wheel everytime these items needed to be protected from
accidential destruction.