Fix: giants could hit incorporeal monsters with anything.
Our hero isn't supposed to be able to hit incorporeal beings (ghosts,
shades, etc) unless their weapon is blessed or enchanted, or is made of
bone or silver. Giants were able to get past this limitation. This
commit fixes that.

Also in this commit, clean up of genus() and add a missing entry for
the infidel quest guardian.