Fix check_wings crash on game restore
`dorecover` calls `restgamestate` then `restlevelstate`. `restlevelstate` does the needed to assign a monster to usteed... meaning for `restgamestate`, `usteed` is gonna be funky. I think that's also partially due to the save & load not storing the actual permonst data, since it's unneeded, and is read back from the original monster in `restlevelstate` anyway.

The issue is that `restgamestate` calls both `check_wings(TRUE)` directly, and also calls `set_uasmon` (which checks wings itself). For both of these, you'll have funky behavior when `check_wings` tries to use the `Flying` macro (which checks `usteed`). Now, the first `check_wings` is removed, leaving only the indirect one, and `set_uasmon` is moved to directly after both the `restgamestate` block and the `restlevelstate` call. This shouldn't break anything, since that function only sets a couple properties & the like, but if intrinsics start vanishing on saving & loading mysterious give that a poke.