Fix samurai quest monsters having odd interactions
The "samurai" on the samurai quest (not the quest guardians, the monsters you fight) were being misnamed and having funky behavior. They weren't getting their rank titles (due to a change in the definition of `is_mplayer`), and they were getting your racial traits due to an overloading of the PM_SAMURAI part. This should fix that.