Fix: missing ranks from enemy samurai during the samurai quest.
This would apply to any player monster, but you only ever see this
happen with the enemy samurai during the samurai quest. The behavior
seen by the same when playing as a giant samurai still exists, and a
partial fix was made (see PR #32, closed) but it's been determined that
a better fix for that would be to just change the enemy monster to
something that the player doesn't use as a role template
(save-breaking).