Colored altars revamp (part 1?)
For awhile now there have been issues with altars either not showing as
the correct color via farlook or when converted, or more seriously, an
altar that started out as aligned and then seemingly at random changes
to unaligned. The devteam actually incorporated colored altars from the
36x-hdf build, but did it their way. Since I have yet to see any odd
altar shenanigans with 3.7.0, I've decided to incorporate it here in the
hopes it clears up these issues. The first issue I listed, I feel
confident this commit addresses that. The latter, we'll see. There are
other more recent commits to 3.7.0 that address other altar/alignment
bits, I may incorporate those as well.