Fix: prevent master mind flayers or alhoons from spawning in the Gnomish Mines.
A fair amount of discussion has occurred about how mind flayers and
master mind flayers can sometimes spawn in the gnomish mines, becasue
'random 'h'' is a thing in the mines' .des file. In EvilHack, because
alhoons are also 'h', they have sometimes (rarely) appeared in the mines
as well. Instead of a blanket change to mkclass and negating all
out-of-depth monsters from forming anywhere, I prefer a more targeted
approach. This commit will prevent alhoons and master mind flayers from
spawning in the mines as those are just way too powerful to ever show
up in that branch, but regular mind flayers are still possible. Can't
take away *all* of the evil...