Prevent certain extraplanar beings from being diseased.
Always found it a bit odd that both angels and demons could get sick.
They are not of the material plane, and are not subject to the ails of
lesser mortal beings :) While looking at this, I noticed that all three
Riders were also not immune to sickness. 'Pestilence looks diseased.
Pestilence  dies from her infection.' - really? The purveyor of all
things diseased and filthy, dying from disease? I think not.