Fix: altars changing from aligned to unaligned when a booby-trapped door exploded.
Another self-induced bug that I introduced several months ago when
correcting another bug dealing with cursed trapped tins and the sokoban
end doors. I had a bit of a mess in b_trapped(), parts of which were
incorrect, causing any aligned altar on that level to flip to unaligned
if a booby-trapped door went off (easy to reproduce in mine town). So
lets clean this up and prevent this bug from occurring anymore. Thanks
to amateurhour and NeroOneTrueKing for the assist in narrowing this
down.