New object property - Excellence.
Addition of a new object property for weapons and armor - Excellence. An
object with this property exhibits this behavior:

* Acts as a luckstone (BUC status applies)
* If armor, its enchantment value adjusts the players charisma attribute
by the same amount when worn (up or down depending on the enchantment
value)
* If a weapon, its BUC status determines the players charisma adjustment
when wielded (25 if blessed, 18 if uncursed, 6 if cursed)

As with all other object properties, it can be wished for on regular
non-magical weapons and armor (wish for <foo> of excellence).

Future tweaks - currently, the luck effects apply when the weapon or
armor with this object property is carried. I'd like to change that so
it only kicks in when the object is worn/wielded. I've done some testing
in this regard, but more work is needed.