Fix: timed clairvoyance, add 'quick_farsight' option.
Added NetHack official commits 48a82fe (timed clairvoyance) and c3fb941
('quick_farsight' option) to address undesireable behavior during the
ascension run with the Amulet of Yendor in inventory. From each commit:

'When the hero has random clairvoyance, the code used
| (moves % 15) == 0 && rn2(2) != 0
(where 'moves' is actually the turn number) to decide when it would
kick in and show a portion of the map.  If the hero was fast enough
to get an extra move when the turn value met the (moves % 15) == 0
condition then clairvoyance could happen twice (or more if poly'd)
on the same turn.'

'Bite the bullet and add a special purpose boolean option to control
game behavior for random clairvoyance.  When objects or monsters are
discovered, it normally issues "you sense your surroundings" and
performs a getpos() operation which allows the player to browse the
map by moving the cursor around and getting 'autodescribe' feedback.
But there have been complaints that once the hero has the Amulet
(which triggers random clairvoyance even though hero isn't flagged
as having that attribute) the message and pause-to-browse become too
intrusive.

This was initially combined with the 'timed clairvoyance' fix because
they both bump EDITLEVEL to invalidate existing save files, but their
details don't interact so I separated them.'

Players will want to toggle 'quick_farsight' in their config if they are
annoyed by the forced random 'you sense your surroundings' messages that
require interction to proceed. With this option set to TRUE,
clairvoyance will just happen at random without any user intervention
being necessary.