Wizard of Yendor buff.
Rodney is supposed to be the most powerful and dangerous wizard in the
entire game, so lets make rumor more like reality. The Wizard of Yendor
can now cast from both schools of monster spells, and has a basic kit
appropriate for a wizard (pointy hat, robe, some type of magical weapon,
chance of some type of amulet).