Fix: crash on reloading game while wielding Grayswandir.
This bug was introduced from an earlier commit (see git commit 061220c)
fixing a crash concerning infidel role and their wings. Moving
set_uasmon() meant youmonst would sometimes return null in certain
cases, such as when wielding grayswandir and checking whether the player
had eyes or not for hallucination checks.

Rather than chase our tails and tack on a fix for that, it would be
better to put set_uasmon() back where it originally belongs, and come up
with a better fix for the issues we were having with check_wings() and
infidels, which would fix both issues. Initial testing passed, but this
is one I want to keep an eye on.

Co-authored-by: RikerW <rikerw@users.noreply.github.com>