New monster spell: reflection.
Higher level spellcasting monsters that can use AD_SPEL can now cast
reflection on themselves, just as the player can with the spell
'reflection'.

This took a lil bit of work to get right. I wound up creating a new MR2
flag for this, essentially making reflection an extrinsic (in this
case). Could be used for other things in the future, but I doubt that
will be necessary as mon_reflects() already handles all others cases of
monsters having reflection.