Fix: most armor special properties were still in effect if it eroded away.
From version 0.7.0:

Before, erode_obj() was only checking for worn gloves. So one could have
had an enchanted helm of brilliance and kept the int/wis benefit if it
had rusted away, or a much more serious case, the alignment change from
wearing a helm of opposite alignment would not revert back if it rusted
away. Now we check for all worn items, not just gloves.