Wand wresting change.
From UnNetHack git commit a763167:

Wand wresting happens always when an empty wand is zapped with a chance
depending on BUC state if an effect is produced. Meaning if you zap an
empty wand in an attempt to wrest a final zap out of it, you have one
chance to wrest a final zap out of it, and its BUC status determines the
chance. If the one chance is unsuccessful, the wand turns to dust. Added
incentive to bless your wands...