Exploding Bag of the Hesperides.
A player recently exploded their Bag of the Hesperides by accident, but
the livelog showed the usual 'just blew up his bag of holding'. This
commit makes it so the livelog event can tell the difference between the
two. Also, an exploding Bag of the Hesperides will do significantly
more damage to the player if it explodes vs an exploding regular bag of
holding - more powerful magic, bigger explosion.