Fix: various artifacts 'incinerating' a monster but they still left a corpse.
Certain artifacts have a small chance to deliver a fatal blow to their
nemesis monster. The feedback for some of those events are things like
'incinerates' or 'bursts into flame' and so on, but sometimes a corpse
would be left behind. The feedback wasn't matching the feedback given.
This commit addresses that.

I also found another bug while working on this - if the player is
attacked by these same artifacts and they are poly'd into that artifacts
nemesis form, the correct event occurs and avoids a panic.