Fix: 'glorkum' from appearing in player monsters inventory.
If the player could not see invisible, and a peaceful player monster
were to spawn, if they were to spawn with a potion of invisibility in
their inventory, it would become a 'glorkum' instead. If the player
interacted with that strange object, it would cause a panic and
potentially crash the game. Thanks qt for catching this one.

Co-authored-by: entrez <me@entrez.cc>