Treat dragon engulfing attack as "swallowing"
Because a dragon will return false from is_animal, and is_animal was
used to test whether something "swallowed" you (i.e. had a stomach, a
mouth, etc) or "engulfed" you (and was a blob or vapor), dragons'
swallowing attacks failed to produce the right kind of messages and
effects in a number of cases.