Fix: material restirctions for sokoban prize objects.
From version 0.7.0:

Discovered it was impossible to find randomly or wish up something like
a mithril helm of speed, even though that would normally be permissible.
The restrictions were put in place to make sure that certain sokoban
prizes would be usable by all playable races without any ill effects. To
resolve this, make sure the sokoban prize flag is enabled with that
restriction - if the flag isn't present, then the full range of object
materials is permissible for those objects.