Hide "magical" status if dknown is false
Previously this was inconsistently done, e.g. an amulet detected from
across the room with a potion of object detection would be identified as
magical, but a scroll would not be.  Make this more consistent -- my
reasoning in eliminating it entirely is that it is more consistent to
withhold information like this when quaffing an uncursed potion of
object detection or stumbling around blind.