Include "magical" for Wiz regardless of item class
There was some inconsistency in-game about whether to show the "magical"
text identifying items as magic only for items that are mundane by
default (e.g. armor) and leave it off for items that are mostly magic
(rings), or whether to show it using the same criteria regardless of
item class (for example, amulets were marked as "magical" even though
all amulets are magic items).

I thought it would be most consistent to just show it for all items,
since even item classes that are magical 'by default' have some
exceptions (meat ring, blank spellbook, paperback book, blank
scroll...), and it is difficult to know where to draw the line with
these, which makes displaying "magical" for some classes and not others
open to debate about exactly which classes should be (in|ex)cluded.

I figure if wizards have an innate sense of what is magical, this sense
operates regardless of item class anyway.