Don't leak identity of soko prize tool
Polyself into a xorn allows the player to phase through the three doors
in the Sokoban treasure zoo and examine each prize up-close.  Using
the #loot command while standing on the 'prize tool' would identify it
as a magic marker ("You don't find anything here to loot") or a bag of
holding ("It's locked"). Similar results for #tip, etc.

I explicitly did not prevent zapping the tool with a wand of probing
providing comparable information; I feel like the other things are
exploits of bugs, but the wand of probing thing seems to me like a
legitimate & creative strategy that is so niche I felt bad about
removing it.