Only print resurrection msg when you can see mon
Also move around where rise_msg is defined for some monsters (those that
are actually *killed* and then regenerated, not those that are merely
"defeated") so that you still get the "You kill it! Maybe not..." when
the special message doesn't print.

In the case of "defeated" monsters, they are still "defeated" so there's
no reason to print "maybe not..." even if you don't see them/get the
more detailed message about their status.