Fix: Illithids can still use psionic blast if poly'd into a non-illithid form.
Illithids/mind flayers can use their psionic powers because of the way
their brains are structured. When polymorphed into a form whose brain is
nowhere near as advanced, that 'ability' should no longer function.

Also in this commit - if an illithid is polymorphed into something else,
they are no longer protected from the effects of amnesia. And give
proper feedback if an illithid tries to use psionic wave but it was
forgotten (which can only happen if they suffer from amnesia while
polymorphed into something else).