Let artifact keys unlock crystal chests
Crystal chests are supposed to be magical chests that are impervious to
ordinary, mundane tools, but artifact unlocking tools (i.e. the Master
Key of Thievery and the PYEC) are not mundane tools since they have
magical artifact powers.  This would be a relatively minor side-benefit
of having one of these artifacts, but I think it makes flavor sense in a
kinda cool way.