Fix: The black dragon eats a black dragon corpse. The black dragon looks very firm.
Technically the feedback the player would see from the commit title was
correct. A black dragon already has disintegration resistance
(mtmp->data/mresists) but not as an intrinsic (mtmp->mintrinsic),
therefore the 'looks very firm' feedback would be seen the first time a
black dragon ate another black dragon. This commit suppresses that
feedback if the monster in question already has that intrinsic via other
means (born with it, or wearing some item giving the extrinsic).

Also a bit of code fixing/clean-up in worn.c